+++ src/main/scala/ch/ethz/queries/chb/Q8.scala
@@ -38,7 +38,7 @@ class Q8 extends ChQuery {
   /**
    * implemented in children classes and hold the actual query
    */
-  override def execute(st: String, cm: String, cn: Int, cs: Int, mUrl: String) = {
+  override def execute(st: String, cm: String, cn: Int, cs: Long, mUrl: String) = {
     val scc = new TSparkContext(mUrl, className, st, cm, cn, cs)
     val sqlContext = new org.apache.spark.sql.SQLContext(scc.sparkContext)
     import org.apache.spark.sql.functions._
